Extract game loop logic from game.tsx into useGameLoop hook
Move physics, bomb movement, collision detection, and game-over effects
into hooks/use-game-loop.ts. game.tsx now only handles UI rendering and
event handling, reducing it from 187 to 68 lines.

Closes #50
